+++ Web/SearchParameters/AspectRatioSearchParameter.cs
@@ -5,24 +5,24 @@ using ImageInfrastructure.Abstractions.Poco;
 namespace ImageInfrastructure.Web.SearchParameters
 {
     public record AspectRatioSearchParameter
-        (NumberComparator Operator, decimal AspectRatio, bool Or = false) : SearchParameter(Or)
+        (NumberComparator Operator, double AspectRatio, bool Or = false) : SearchParameter(Or)
     {
         public override Expression<Func<Image, bool>> Evaluate()
         {
             switch (Operator)
             {
                 case NumberComparator.NotEqual:
-                    return a => Math.Abs((decimal) a.Width / a.Height - AspectRatio) > 0.01M;
+                    return a => Math.Abs((double)a.Width / a.Height - AspectRatio) > 0.01D;
                 case NumberComparator.Equal:
-                    return a => Math.Abs((decimal) a.Width / a.Height - AspectRatio) < 0.01M;
+                    return a => Math.Abs((double)a.Width / a.Height - AspectRatio) < 0.01D;
                 case NumberComparator.GreaterThan:
-                    return a => (decimal) a.Width / a.Height > AspectRatio;
+                    return a => (double)a.Width / a.Height > AspectRatio;
                 case NumberComparator.LessThan:
-                    return a => (decimal) a.Width / a.Height < AspectRatio;
+                    return a => (double)a.Width / a.Height < AspectRatio;
                 case NumberComparator.GreaterThan | NumberComparator.Equal:
-                    return a => (decimal) a.Width / a.Height > AspectRatio || Math.Abs((decimal) a.Width / a.Height - AspectRatio) < 0.01M;
+                    return a => (double)a.Width / a.Height > AspectRatio || Math.Abs((double)a.Width / a.Height - AspectRatio) < 0.01D;
                 case NumberComparator.LessThan | NumberComparator.Equal:
-                    return a => (decimal) a.Width / a.Height < AspectRatio || Math.Abs((decimal) a.Width / a.Height - AspectRatio) < 0.01M;
+                    return a => (double)a.Width / a.Height < AspectRatio || Math.Abs((double)a.Width / a.Height - AspectRatio) < 0.01D;
                 default:
                     throw new ArgumentOutOfRangeException();
             }
